refactor(devtools): rename fetch.ts->state.ts, explicit layer imports, bump ^5.2.4
- lib/schema-org/fetch.ts -> state.ts (naming convention)
- use explicit nuxtseo-layer-devtools/composables/* imports (not #imports)
- bump nuxtseo-layer-devtools/nuxtseo-shared -> ^5.2.4, bare release-age excludes
