+++ pnpm-workspace.yaml
@@ -6,8 +6,8 @@ minimumReleaseAgeExclude:
   - unhead@3.1.3
   - '@unhead/bundler@3.1.3'
   - '@unhead/vue@3.1.3'
-  - nuxtseo-layer-devtools@5.2.2
-  - nuxtseo-shared@5.2.2
+  - nuxtseo-layer-devtools
+  - nuxtseo-shared
 shellEmulator: true
 
 trustPolicy: no-downgrade
@@ -43,8 +43,8 @@ catalog:
   happy-dom: ^20.10.2
   nuxt: ^4.4.8
   nuxt-site-config: ^4.0.8
-  nuxtseo-layer-devtools: ^5.2.2
-  nuxtseo-shared: ^5.2.2
+  nuxtseo-layer-devtools: ^5.2.4
+  nuxtseo-shared: ^5.2.4
   pkg-types: ^2.3.1
   typescript: ^6.0.3
   ufo: ^1.6.4
